ai-app-processing-service/requirements.txt
psycopg2-binary
python-dotenv
pydantic
